Fixed an important issue After getting and printing the control_list self.data from GuildData was breaking. Instead of copying the control_list to a new array, I was creating a reference to it and then changing it later. Now it copies the list and doesn't touch the original one.
Also changed how the permissions are when it joins a server. The control_list is empty
and everyone can control it. everyone is no longer a role (that didn't work well) so
you have to add at least 1 role to the list or everyone will be able to control it.