Changed how the bot gets new announcements
Instead of searching for differences between the new and old rss feed, now it checks for announcements that are newer than a timestamp. Instead of storing all of the rss feed in a file, now it only stores the last timestamp. It still needs testing to make sure there are no missing announcements.